use rune type instead of int for unicode tables
int is 64 bit on 64 bit platforms but each unicode char is only 32bit
and go APIs generally use rune as type for it so switch to that.

The main motivation is to reduce the binary size with a small example
program I see a about 170k reduce binary size footprint.

Note I went with minimal changes in terms of changing the API of various
internal functions and cast back to int quite early. In theory one could
make much more changes to use rune over int.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>